Refractor and Import Assets
Changed the SFX structure so that clips can be linked to a enum and be played that way. This allows for easier set up in inspector.
Also imported lots of assets.

P.S. The other option I've considered is the approach of using strings to identify clips. Considering the annoyance that using strings have caused from other systems, I think it's worth exploring enums as an option to simplify the process.