Fixed camera error
The problem was that the camera zoom updated every time the conversation started or ended.  But this should only happen if the concept tab is actually enabled.  So we have to do a check to make sure the canvas game object is active in the heirarchy first, before updating camera zoom