Changed the template
The template has a list of objects. Each object is either a single question or a pool of questions. The main issue with this architecture is that all the previous data would not be serialized correctly, resulting in loss of data from all quiz scriptable objects.  It might be better to change the architecture so that there is only one list of questions and they are either all randomized or they aren't.  Should align with James the writer