Preparing to make conversation repeatable
Packaged the code that says the quiz conversation into a little function so that it will be easier to say that conversation again after failing the quiz.  But I must be missing some code here because the quiz conversation response should not automatically load the next level.  Will need to pull staging before continuing