Fixed the food update problem
For some reason the sorted set removed one of the populations from the list. The result was that one of the populations was not having its food need updated. The new system creates a new list and sorts it without using a sorted set. Not sure why the sorted set would have done that though