Animal moving bug fixed
Animals can no longer be placed in the middle of nowhere, and can only be placed on traversible terrain.  However, there is a new problem.  It appears that transplanting an animal to an area that is cut off from its original population results in a population split, where the new population is permanently stagnate.  This probably has a similar root cause to the poulation split bug discovered recently where the food needs of one of the split populations was not being updated