Improved bookmark data
Before the bookmark data was not serializable so that there were null reference exceptions when reading it from a file.  The data has now been converted to serializable and uses Reflection to determine which data fields to set in a way that "looks" generic to other classes that want to set the data.  The design is messy-ish but could be a lot worse.  The only other solution would be to make any notebook data inherit from the same base class, which is also not the best solution because those classes may want to inherit from their own classes that are not related to the bookmarking system