Population no longer drops to zero
Before, the population would reduce the number of animals if the animal was in an enclosed area where isEnclosed returned false.  There is no obvious reason why this should be the case.  The structure of the code still exists but the part of the code that reduces the number of animals has been simply removed.  I would think that this could cause some unforseen issues, but some basic testing with population splitting / combining has not resulted in any problems