Need changed to use item ids
No runtime errors related to the item id refactor have been detected yet.  What I'm most concerned about would be the FoodSourceSystem interacting with the needs in the Population.  The needs in the population have to map strings to needs because some needs have names that are not associated with any item in the item registry.  This should probably be refactored, but the roots of the string comparison grow very deep, all the way into the liquid need systems