Changed scriptable object singleton
This used to be using a "LoadAll" command.  But that was creating an uncomfortable lag in the editor whenever different classes wanted to load the singleton.  Now, the singleton has to be in the root of a resources folder with the correct name, but it'll be more efficient and throw a detailed exception if it fails