Allowed terrain computation
The terrain need system should take water into account because kelp needs to be placed in water for its terrain need.  There is currently a conflict between the terrain needs and water needs of the Kelp.  It needs water as a terrain need AND it needs water as a liquid need.  This is a problem because the need values for the food source has to map an item ID to a need, and we cannot use the same id for the same key.  The way this was worked around before was that the dictionary mapped strings to the need, and the string key was slightly different for the "water as terrain" need and the "water as liquid" need.  It's not at all clear yet how this can be resolved, because lots of classes are expecting the "Life" interface to give one complete dictionary with all the needs.  I would have liked to try splitting up all the needs into prey, food, terrain, and liquid but the design for one single dictionary with all the needs is depended on by lots and lots of other classes