Created temporary workaround for water terrain clash
The problem was that we could not have the same item id in the dictionary or needs for a Life object.  This is circumvented using a special standalone terrain water need.  This only works for food sources right now.  Next, we have to get it to work for species because slugs also need water as a terrain need