Trailblazing Momos
The need rating builder can now compute the tree needs and friend needs of the species.  It is not clear yet whether the computation is exactly what the designers want but it works really well regardless.  It just reuses the same computation from food and terrain and works really well.  Also, the way the need sliders are displayed has changed